requirements.txt
datasets
tqdm
Pillow
huggingface_hub
